PVS-Studio fix pass: 10 parallel agents, ~70 files of production-code fixes
Agent 4 (V501 x996 false positives): single-line consteval helper in
tdutils/td/utils/logging.h replaces the `integral_constant<int,A>() >
integral_constant<int,A>()` tautology with a named predicate call —
silences the entire 996-finding class at the source.

Agent 6 (V1051 x29 misprints): all 29 instances confirmed false
positives — standard move-then-check, parameter-to-member copy, and
structurally-similar-but-distinct variable pairs. No changes.

Agent 10 (V678 x66 actor_id): confirmed false positive — TDLib actor
framework idiom where `Actor::actor_id(this)` uses its pointer
argument solely for SelfT template type deduction, not as a genuine
self-argument. No changes.

Agents 1, 2, 3, 5, 7, 8, 9: hit their usage caps partway through but
produced fixes across ~70 production files:
- V730 (uninitialized class members): default member initializers
  added in td/telegram/ managers and header-declared nested structs
  (BackgroundManager, ChatManager, GroupCallManager, MessagesManager,
  MessageEntity, StoryManager, UserManager, UpdatesManager, etc.)
  and td/mtproto/ (AuthData, StealthConfig).
- V547 / V560 (dead conditions): removed provably-true guards in
  td/telegram/* and td/mtproto/stealth/StealthTransportDecorator.
- V530 (unused return values): explicit (void) casts on
  to_lower_inplace() calls across LanguagePackManager, LinkManager,
  misc, HttpReader.
- V1048 (same-value assignments): removed redundant assigns in
  BrowserProfile factory helpers where the struct header already
  defaults the field (TlsVersion::Tls12 was the dominant case).
- V570, V590, V607, V654, V763 (misc HIGH): addressed in test helpers
  and ConcurrentHashTable.h.
- V1028 (overflow before cast): test helpers cast operands to size_t
  before arithmetic instead of casting the result.

Pre-existing build fix: WireClassifierFeatures.cpp used
std::filesystem::path::native() in a StringBuilder << context — fails
on Windows because native() returns std::wstring which td::StringBuilder
doesn't accept. Replaced with path::string(). Root cause was
upstream commit 28e08dca9 ("harden ssl trust store"); not PVS-related.

Build: 271/271 compile units green on Windows MSVC Debug + vcpkg
OpenSSL.

Tests: all new workstream tests green (SEKV §6.11 71/71, Workstream B
36/36, WS-C 11/11, WS-D 17/17, WS-E 15/15, WS-F 18/18 PR-scope,
StealthParamsLoader 23/23, BidirectionalSize/IdleChaff/RecordPadding/
StreamTransport/Session*/RawConnection/Handshake/Ping 179/179). One
pre-existing upstream UNREACHABLE in
Test_StealthConfigProfileRecordLimit_TlsSecretsUseRuntimeProfileSelectionKey
— reproduces without these changes via git stash; out of scope for
this commit, tracked separately.

Analysis_Report.json added to .gitignore.

+++ test/message_entities.cpp
@@ -863,14 +863,14 @@ TEST(MessageEntities, fix_formatted_text) {
   str = "aba \r\n caba ";
   td::UserId user_id(static_cast<td::int64>(1));
   for (td::int32 length = 1; length <= 3; length++) {
-    for (td::int32 offset = 0; static_cast<size_t>(offset + length) <= str.size(); offset++) {
+    for (td::int32 offset = 0; static_cast<size_t>(offset) + static_cast<size_t>(length) <= str.size(); offset++) {
       for (auto type : {td::MessageEntity::Type::Bold, td::MessageEntity::Type::Url, td::MessageEntity::Type::TextUrl,
                         td::MessageEntity::Type::MentionName}) {
         for (auto skip_trim : {false, true}) {
           fixed_str = skip_trim ? "aba \n caba " : "aba \n caba";
           auto fixed_length = offset <= 4 && offset + length >= 5 ? length - 1 : length;
           auto fixed_offset = offset >= 5 ? offset - 1 : offset;
-          while (static_cast<size_t>(fixed_offset + fixed_length) > fixed_str.size()) {
+          while (static_cast<size_t>(fixed_offset) + static_cast<size_t>(fixed_length) > fixed_str.size()) {
             fixed_length--;
           }
 
@@ -901,7 +901,7 @@ TEST(MessageEntities, fix_formatted_text) {
     for (td::int32 offset = -10; offset <= 10; offset++) {
       td::vector<td::MessageEntity> entities;
       entities.emplace_back(td::MessageEntity::Type::Bold, offset, length);
-      if (length < 0 || offset < 0 || (length > 0 && static_cast<size_t>(length + offset) > str.size())) {
+      if (length < 0 || offset < 0 || (length > 0 && static_cast<size_t>(length) + static_cast<size_t>(offset) > str.size())) {
         check_fix_formatted_text(str, entities, true, false, false, false);
         check_fix_formatted_text(str, entities, false, false, false, true);
         continue;
